Adds remove stale endpoints script
